Walking tool restructure and refactor

+++ data/tools/walking/src/main/java/fit/asta/health/data/walking/service/StepCounterState.kt
@@ -6,7 +6,7 @@ data class StepCounterState(
     val date: LocalDate,
     val steps: Int,
     val goalDistance: Float,
-    val goalDuration: Int,
+    val goalDuration: Float,
     val distanceTravelled: Double,
     val calorieBurned: Int,
     val duration: Int
